requirements.txt
django==4.0.2
djangorestframework==3.13.1
django-cors-headers==3.11.0
psycopg2-binary==2.9.3
pyyaml>=6.0
uwsgi>=2.0
